Moved some functions to a shared crate
