+++ examples/runners/ash/Cargo.toml
@@ -13,6 +13,7 @@ use-installed-tools = ["spirv-builder/use-installed-tools"]
 use-compiled-tools = ["spirv-builder/use-compiled-tools"]
 
 [dependencies]
+shared = { path = "../../shaders/shared" }
 ash = "0.31"
 ash-window = "0.5"
 cfg-if = "1.0.0"
