+++ examples/runners/cpu/Cargo.toml
@@ -9,6 +9,7 @@ publish = false
 [dependencies]
 minifb = "0.19.1"
 # bring in the shader as natively compiled code
+shared = { path = "../../shaders/shared" }
 sky-shader = { path = "../../shaders/sky-shader" }
 spirv-std = { path = "../../../crates/spirv-std" }
