+++ examples/runners/wgpu/Cargo.toml
@@ -17,6 +17,7 @@ use-compiled-tools = ["spirv-builder/use-compiled-tools"]
 
 [dependencies]
 cfg-if = "1.0.0"
+shared = { path = "../../shaders/shared" }
 futures = { version = "0.3", default-features = false, features = ["std", "executor"] }
 wgpu = "0.6.0"
 winit = { version = "0.23", features = ["web-sys"] }
