+++ examples/shaders/shared/Cargo.toml
@@ -1,14 +1,10 @@
 [package]
-name = "sky-shader"
+name = "shared"
 version = "0.1.0"
 authors = ["Embark <opensource@embark-studios.com>"]
 edition = "2018"
 license = "MIT OR Apache-2.0"
 publish = false
 
-[lib]
-crate-type = ["dylib"]
-
 [dependencies]
-shared = { path = "../../shaders/shared" }
 spirv-std = { path = "../../../crates/spirv-std" }
