+++ examples/shaders/sky-shader/src/lib.rs
@@ -6,6 +6,7 @@
 #![register_attr(spirv)]
 
 use core::f32::consts::PI;
+use shared::*;
 use spirv_std::glam::{const_vec3, Vec2, Vec3, Vec4};
 use spirv_std::{Input, MathExt, Output, PushConstant};
 
@@ -30,37 +31,6 @@ pub struct ShaderConstants {
     pub width: u32,
     pub height: u32,
     pub time: f32,
-}
-
-// TODO: add this to glam? Rust std has it on f32/f64
-fn pow(v: Vec3, power: f32) -> Vec3 {
-    Vec3::new(v.x().pow(power), v.y().pow(power), v.z().pow(power))
-}
-
-// TODO: add this to glam? Rust std has it on f32/f64
-fn exp(v: Vec3) -> Vec3 {
-    Vec3::new(v.x().exp(), v.y().exp(), v.z().exp())
-}
-
-/// Based on: https://seblagarde.wordpress.com/2014/12/01/inverse-trigonometric-functions-gpu-optimization-for-amd-gcn-architecture/
-fn acos_approx(v: f32) -> f32 {
-    let x = v.abs();
-    let mut res = -0.155972 * x + 1.56467; // p(x)
-    res *= (1.0f32 - x).sqrt();
-
-    if v >= 0.0 {
-        res
-    } else {
-        PI - res
-    }
-}
-
-/// renamed because of cross-compilation issues with spirv-cross/ moltenvk
-fn my_smoothstep(edge0: f32, edge1: f32, x: f32) -> f32 {
-    // Scale, bias and saturate x to 0..1 range
-    let x = ((x - edge0) / (edge1 - edge0)).saturate();
-    // Evaluate polynomial
-    x * x * (3.0 - 2.0 * x)
 }
 
 fn total_rayleigh(lambda: Vec3) -> Vec3 {
@@ -91,17 +61,6 @@ fn sun_intensity(zenith_angle_cos: f32) -> f32 {
                 / SUN_INTENSITY_FALLOFF_STEEPNESS))
                 .exp(),
         )
-}
-
-fn tonemap(col: Vec3) -> Vec3 {
-    // see https://www.desmos.com/calculator/0eo9pzo1at
-    const A: f32 = 2.35;
-    const B: f32 = 2.8826666;
-    const C: f32 = 789.7459;
-    const D: f32 = 0.935;
-
-    let z = pow(col, A);
-    z / (pow(z, D) * B + Vec3::splat(C))
 }
 
 fn sky(dir: Vec3, sun_position: Vec3) -> Vec3 {
@@ -154,13 +113,6 @@ fn sky(dir: Vec3, sun_position: Vec3) -> Vec3 {
     l0 += sun_e * 19000.0 * fex * sundisk;
 
     lin + l0
-}
-
-fn get_ray_dir(uv: Vec2, pos: Vec3, look_at_pos: Vec3) -> Vec3 {
-    let forward = (look_at_pos - pos).normalize();
-    let right = Vec3::new(0.0, 1.0, 0.0).cross(forward).normalize();
-    let up = forward.cross(right);
-    (forward + uv.x() * right + uv.y() * up).normalize()
 }
 
 pub fn fs(constants: &ShaderConstants, frag_coord: Vec2) -> Vec4 {
